Add files via upload
these are updated files, mainly fixes for spacing on admin_poll_add.dart and poll_results.dart. the theme is not implemented yet but should be done sometime late tonight.

there are still spacing issues on other pages, like the login page (top is too close to the top edge of the phone emulator). the settings page also takes you back to login.